BanDongHanh_Website/requirements-min.txt
# ================== MINIMAL REQUIREMENTS FOR CHAT FUNCTIONALITY ==================
# Essential dependencies to run the core chat feature with basic functionality

# Core web framework
streamlit>=1.28.0              # Main application framework

# AI integration - choose one based on your setup
openai>=1.3.0                  # OpenAI GPT models (primary)
# google-generativeai>=0.3.0   # Alternative: Google Gemini (uncomment if using)

# Essential utilities
Pillow>=9.0.0                  # Image processing for image chat support
python-dotenv>=0.19.0          # Environment variables (.env file support)
requests>=2.28.0               # HTTP requests for API communication

# Note: This minimal setup provides core chat functionality.
# For full features (TTS, advanced memory, cloud services), install requirements.txt